broken - working on packer AMI for controllers.
Currently working on getting the "initial cluster" param values to match
the "name" param value. The kubernetes the hard way tutorial assumes the
hostname matches the instance name, but mine does not. I need to find a
way to either get the hostname to match the instance name, or collect
the "Name" tag for the instance and expose it to the template when it is
being written.

Once etcd is running properly, need to convert the Terraform to use the
controller AMI.